+++ Models/basic/Profile.cs
@@ -6,18 +6,18 @@ public partial class Profile
     {
     }
 
-    public long ProfileId {get; set;}
-    public long UserId {get; set;}
+    public int ProfileId {get; set;}
+    public int UserId {get; set;}
     public required string Name {get; set;}
     public string? Photo {get; set;}
     public required string DateOfBirth {get; set;} // made string for now to avoid potential server conflicts
 
-    public IList<long>? SubtitleSettings {get; set;}
-    public IList<long>? ContentTypePreferences {get; set;}
-    public IList<long>? GenrePreferences {get; set;}
-    public IList<long>? ClassificationPreferences {get; set;}
-    public IList<long>? WantToWatch {get; set;}
-    public IList<long>? ContentSession {get; set;}
+    public IList<int>? SubtitleSettings {get; set;}
+    public IList<int>? ContentTypePreferences {get; set;}
+    public IList<int>? GenrePreferences {get; set;}
+    public IList<int>? ClassificationPreferences {get; set;}
+    public IList<int>? WantToWatch {get; set;}
+    public IList<int>? ContentSession {get; set;}
 
     
 }
