+++ Models/basic/TextItem.cs
@@ -6,8 +6,8 @@ public partial class TextItem
     {
     }
 
-    public long Id {get; set;}
+    public int Id {get; set;}
     public required string OnPageTextId {get; set;}
 
-    public virtual required IList<long> Languages {get; set;}
+    public virtual required IList<int> Languages {get; set;}
 }
