+++ Models/manyToMany/EpisodeXQualityAvailable.cs
@@ -5,7 +5,7 @@ namespace NetflixAPI.Models;
 [PrimaryKey(nameof(EpisodeId), nameof(QualityId))]
 public class EpisodeXQualityAvailable
 {
-    public long EpisodeId { get; set; }
-    public long QualityId { get; set; }
+    public int EpisodeId { get; set; }
+    public int QualityId { get; set; }
 
 }
