+++ Models/manyToMany/SubtitleSettings.cs
@@ -10,9 +10,9 @@ public partial class SubtitleSettings
 
     }
 
-    public required long Size {get; set;}
+    public required int Size {get; set;}
     public required string Color {get; set;}
-    public long ProfileId {get; set;}
-    public long LanguageId {get; set;}
+    public int ProfileId {get; set;}
+    public int LanguageId {get; set;}
 
 }
